database+SQL
Board Games
Felicia stocks many games, all games have a name, one or more publisher, a publication date, a designer, a price, and the UPC (sample data does not need to be accurate). She will also need to record the number of copies of a game she has in stock.
Memberships
Memberships are free! When a customer signs up, they get a card with their membership number. Felicia wants to track the number of games they’ve bought. She also wants to keep the customer’s name and email address, so that she can let members know about upcoming deals.
Sales/Receipts
Not Board Anymore sales are tracked, so that Felicia can help customers with returns, and get sales reports and see how her business is doing. When a sale is made each product purchased is recorded which product(s) (each sale can include many products), were purchased, and the time/date of the transaction need to be stored. Additionally, if a member purchases a game, their membership is recorded.
Task
Create a relational database structure to support Felicia’s current needs. 
Create a relationship schema diagram, and MariaDB database with the tables and some sample data, so that the structure can be verified.
Add the ability to store “related” games, a game record can reference any number of related games and could be related for different reasons

Create views that:
1. Show total number and amount ($) of sales in the past month
2. Shows games with fewer than 3 copies in stock (2 or less)
3. Shows total number and amount ($) of sales in the past year to non-members
4. Shows total number and amount ($) of sales in the past year to members
5. Show View 3 and 4 (above) in a single view.
Queries
1. Find all the games by a publisher with “man” anywhere in the name of the publisher
2. Updates the number of games a specific member has purchased (any number will do)
3. Removes a member from the database
4. Adds a new game